feat(demo): extract shared utilities to vultron/demo/utils.py (DEMO-4.1)
Create vultron/demo/utils.py with demo_step, demo_check context managers
and HTTP client helpers (DataLayerClient, logfmt, postfmt, reset_datalayer,
discover_actors, init_actor_ios, post_to_inbox_and_wait, verify_object_stored,
get_offer_from_datalayer, log_case_state, setup_clean_environment,
check_server_availability, BASE_URL) extracted from initialize_case_demo.py.

Update initialize_case_demo.py to import from vultron.demo.utils instead
of defining utilities locally. All other demo scripts continue to work via
their existing imports from initialize_case_demo.

Satisfies DC-02-001 (partial): from vultron.demo.utils import demo_step,
demo_check succeeds. Full completion (all scripts importing directly) is
DEMO-4.2.
